requirements.txt
cryptography
python-jose[cryptography]
passlib[bcrypt]
fastapi
fastapi-jwt
sqlalchemy
alembic
dotenv
psycopg2-binary
uvicorn
asyncpg
pydantic-extra-types
phonenumbers
python-multipart
boto3
pandas
numpy
beautifulsoup4==4.12.3
loguru==0.7.0
notifiers==1.3.3
python-dotenv==1.0.1
requests==2.32.3
seleniumbase
openpyxl
paramiko==3.4.0
